# Conflicts:
#	.gitignore
#	README.rst
#	askbot/__init__.py
#	askbot/conf/__init__.py
#	askbot/deps/django_authopenid/forms.py
#	askbot/deps/django_authopenid/urls.py
#	askbot/deps/django_authopenid/views.py
#	askbot/deps/group_messaging/views.py
#	askbot/doc/source/changelog.rst
#	askbot/doc/source/index.rst
#	askbot/feed.py
#	askbot/forms.py
#	askbot/locale/ar/LC_MESSAGES/django.mo
#	askbot/locale/ar/LC_MESSAGES/django.po
#	askbot/locale/ar/LC_MESSAGES/djangojs.mo
#	askbot/locale/ar/LC_MESSAGES/djangojs.po
#	askbot/locale/de/LC_MESSAGES/django.mo
#	askbot/locale/de/LC_MESSAGES/django.po
#	askbot/locale/en/LC_MESSAGES/django.po
#	askbot/locale/es/LC_MESSAGES/django.mo
#	askbot/locale/es/LC_MESSAGES/django.po
#	askbot/locale/fr/LC_MESSAGES/django.mo
#	askbot/locale/fr/LC_MESSAGES/django.po
#	askbot/locale/fr/LC_MESSAGES/djangojs.po
#	askbot/locale/it/LC_MESSAGES/django.mo
#	askbot/locale/it/LC_MESSAGES/django.po
#	askbot/locale/it/LC_MESSAGES/djangojs.mo
#	askbot/locale/it/LC_MESSAGES/djangojs.po
#	askbot/locale/ko/LC_MESSAGES/django.po
#	askbot/locale/ko/LC_MESSAGES/djangojs.mo
#	askbot/locale/ko/LC_MESSAGES/djangojs.po
#	askbot/locale/nb_NO/LC_MESSAGES/django.mo
#	askbot/locale/nb_NO/LC_MESSAGES/django.po
#	askbot/locale/vi/LC_MESSAGES/django.mo
#	askbot/locale/vi/LC_MESSAGES/django.po
#	askbot/locale/vi/LC_MESSAGES/djangojs.mo
#	askbot/locale/vi/LC_MESSAGES/djangojs.po
#	askbot/locale/zh_CN/LC_MESSAGES/django.mo
#	askbot/locale/zh_CN/LC_MESSAGES/django.po
#	askbot/locale/zh_HK/LC_MESSAGES/django.mo
#	askbot/locale/zh_HK/LC_MESSAGES/django.po
#	askbot/locale/zh_TW/LC_MESSAGES/django.mo
#	askbot/locale/zh_TW/LC_MESSAGES/django.po
#	askbot/locale/zh_TW/LC_MESSAGES/djangojs.mo
#	askbot/locale/zh_TW/LC_MESSAGES/djangojs.po
#	askbot/mail/__init__.py
#	askbot/mail/messages.py
#	askbot/mail/parsing.py
#	askbot/management/commands/askbot_import_jive.py
#	askbot/management/commands/generate_post_snippets.py
#	askbot/management/commands/merge_users.py
#	askbot/management/commands/send_email_alerts.py
#	askbot/media/jquery-openid/jquery.openid.js
#	askbot/media/jquery-openid/openid.css
#	askbot/media/js/live_search.js
#	askbot/media/js/post.js
#	askbot/media/js/user.js
#	askbot/media/js/utils.js
#	askbot/media/style/style.css
#	askbot/media/style/style.less
#	askbot/models/__init__.py
#	askbot/models/post.py
#	askbot/models/question.py
#	askbot/search/haystack/__init__.py
#	askbot/search/state_manager.py
#	askbot/skins/loaders.py
#	askbot/templates/answer_edit.html
#	askbot/templates/ask.html
#	askbot/templates/authopenid/complete.html
#	askbot/templates/authopenid/providers_javascript.html
#	askbot/templates/authopenid/signin.html
#	askbot/templates/authopenid/widget_signin.html
#	askbot/templates/email/delayed_email_alert.html
#	askbot/templates/email/feedback_email.txt
#	askbot/templates/email/macros.html
#	askbot/templates/email/notify_admins_about_new_tags.html
#	askbot/templates/embed/ask_by_widget.html
#	askbot/templates/feedback.html
#	askbot/templates/macros.html
#	askbot/templates/main_page/headline.html
#	askbot/templates/main_page/javascript.html
#	askbot/templates/main_page/questions_loop.html
#	askbot/templates/meta/bottom_scripts.html
#	askbot/templates/meta/fonts.html
#	askbot/templates/meta/html_head_javascript.html
#	askbot/templates/question.html
#	askbot/templates/question/answer_card.html
#	askbot/templates/question/closed_question_info.html
#	askbot/templates/question/content.html
#	askbot/templates/question/javascript.html
#	askbot/templates/question/new_answer_form.html
#	askbot/templates/question/question_card.html
#	askbot/templates/question/sidebar.html
#	askbot/templates/subscribe_for_tags.html
#	askbot/templates/tags.html
#	askbot/templates/tags/list_bulk_tag_subscription.html
#	askbot/templates/user_profile/user.html
#	askbot/templates/user_profile/user_answers_list.html
#	askbot/templates/user_profile/user_edit.html
#	askbot/templates/user_profile/user_email_subscriptions.html
#	askbot/templates/user_profile/user_info.html
#	askbot/templates/user_profile/user_reputation.html
#	askbot/templates/user_profile/user_stats.html
#	askbot/templates/user_profile/user_tabs.html
#	askbot/templates/users.html
#	askbot/templates/widgets/ask_button.html
#	askbot/templates/widgets/ask_form.html
#	askbot/templates/widgets/edit_post.html
#	askbot/templates/widgets/logo.html
#	askbot/templates/widgets/question_summary.html
#	askbot/templates/widgets/related_tags.html
#	askbot/templates/widgets/scope_nav.html
#	askbot/templates/widgets/user_navigation.html
#	askbot/tests/email_parsing_tests.py
#	askbot/tests/jive_tests.py
#	askbot/urls.py
#	askbot/utils/forms.py
#	askbot/utils/html.py
#	askbot/utils/jive.py
#	askbot/utils/transaction.py
#	askbot/views/api_v1.py
#	askbot/views/avatar_views.py
#	askbot/views/readers.py
#	askbot/views/users.py
#	askbot/views/writers.py
#	askbot_requirements.txt
#	askbot_requirements_dev.txt

+++ askbot/management/commands/delete_unused_tags.py
@@ -6,7 +6,6 @@ from askbot.conf import settings as askbot_settings
 import sys
 
 class Command(NoArgsCommand):
-    @transaction.commit_manually
     def handle_noargs(self, **options):
         tags = models.Tag.objects.all()
         message = 'Searching for unused tags:'
@@ -15,9 +14,18 @@ class Command(NoArgsCommand):
         deleted_tags = list()
         for tag in ProgressBar(tags, total, message):
             if not tag.threads.exists():
-                deleted_tags.append(tag.name)
-                tag.delete()
-            transaction.commit()
+                #if any user subscribed for the tag and
+                #the user is not blocked, skip deleting the tag
+                marks = tag.user_selections.all()
+                do_delete = True
+                for mark in marks:
+                    if not mark.user.is_blocked():
+                        do_delete = False
+                        break
+
+                if do_delete:
+                    deleted_tags.append(tag.name)
+                    tag.delete()
 
         if deleted_tags:
             found_count = len(deleted_tags)
